Listening onSnapshot in ScoreStore to save any change in scores node directly. Moving Store to a new stores directory.
